Implemented float tensor with documentation and testing

+++ csalt/c_int16.c
@@ -19,7 +19,7 @@
 // ================================================================================ 
 // ================================================================================ 
 
-int16_tensor_expect_t init_int16_tensor(int16_t            ndim,
+int16_tensor_expect_t init_int16_tensor(uint8_t            ndim,
                                         const size_t*      shape,
                                         allocator_vtable_t alloc_v) {
     /* Allocate the underlying tensor */
